fix: rename setupTest
